Extract fixtures for convention catalog tests
