+++ src/test/kotlin/dev/panuszewski/gradle/ConventionPluginsSpec.kt
@@ -1,13 +1,16 @@
 package dev.panuszewski.gradle
 
 import dev.panuszewski.gradle.fixtures.CommentedPluginUsage
+import dev.panuszewski.gradle.fixtures.ConventionCatalogUsedFromConventionPlugin
+import dev.panuszewski.gradle.fixtures.ConventionCatalogUsedInParentBuild
+import dev.panuszewski.gradle.fixtures.ConventionCatalogUsedInParentBuildThatIsNotRootBuild
+import dev.panuszewski.gradle.fixtures.ConventionCatalogUsedInRootBuildThatIsNotDirectParent
 import dev.panuszewski.gradle.fixtures.CustomBuildDirPath
 import dev.panuszewski.gradle.fixtures.ImportedCatalog
 import dev.panuszewski.gradle.fixtures.LibsInDependenciesBlock
 import dev.panuszewski.gradle.fixtures.LibsInPluginsBlock
 import dev.panuszewski.gradle.fixtures.LibsInPluginsBlockInCustomLocation
 import dev.panuszewski.gradle.fixtures.MultiLevelBuildHierarchy
-import dev.panuszewski.gradle.fixtures.MultiLevelBuildHierarchy.someLibrary
 import dev.panuszewski.gradle.fixtures.MultipleCatalogsInDependenciesBlock
 import dev.panuszewski.gradle.fixtures.MultipleCatalogsInPluginsBlock
 import dev.panuszewski.gradle.fixtures.MultipleConventionPlugins
@@ -17,7 +20,6 @@ import dev.panuszewski.gradle.fixtures.TypesafeConventionsAppliedToIncludedBuild
 import dev.panuszewski.gradle.fixtures.TypesafeConventionsConfig
 import dev.panuszewski.gradle.fixtures.includedbuild.BuildLogic
 import dev.panuszewski.gradle.fixtures.includedbuild.PluginManagementBuildLogic
-import dev.panuszewski.gradle.framework.BuildOutcome
 import dev.panuszewski.gradle.framework.BuildOutcome.BUILD_FAILED
 import dev.panuszewski.gradle.framework.BuildOutcome.BUILD_SUCCESSFUL
 import dev.panuszewski.gradle.framework.Fixture
@@ -352,41 +354,7 @@ class ConventionPluginsSpec : GradleSpec() {
     fun `should generate typesafe accessor for convention plugin and use it from parent build`() {
         // given
         installFixture(BuildLogic)
-        installFixture(TypesafeConventionsAppliedToIncludedBuild)
-
-        buildGradleKts {
-            """
-            plugins {
-                alias(conventions.plugins.someConvention)
-            }
-            
-            repositories {
-                mavenCentral()
-            }
-            """
-        }
-
-        includedBuild {
-            buildGradleKts {
-                """
-                plugins {
-                    `kotlin-dsl`
-                }
-                
-                repositories {
-                    gradlePluginPortal()
-                }
-                """
-            }
-
-            customProjectFile("src/main/kotlin/conventions/someConvention.gradle.kts") {
-                """
-                package conventions
-                    
-                println("Hello from someConvention")    
-                """
-            }
-        }
+        installFixture(ConventionCatalogUsedInParentBuild)
 
         // when
         val result = runGradle()
@@ -399,84 +367,7 @@ class ConventionPluginsSpec : GradleSpec() {
     @Test
     fun `should generate typesafe accessor for convention plugin and use it from parent build in multi-level hierarchy`() {
         // given
-        val secondaryBuild = mainBuild.registerIncludedBuild("secondary-build")
-        val buildLogic = secondaryBuild.registerIncludedBuild("build-logic")
-
-        with(mainBuild) {
-            buildGradleKts {
-                """
-                plugins {
-                    java
-                }
-                """
-            }
-        }
-
-        with(secondaryBuild) {
-            libsVersionsToml {
-                """
-                [libraries]
-                some-library = "$someLibrary"    
-                """
-            }
-
-            buildGradleKts {
-                """
-                plugins {
-                    alias(conventions.plugins.someConvention)
-                }
-                
-                group = "com.example"
-                
-                repositories {
-                    mavenCentral()
-                }
-                """
-            }
-        }
-
-        with(buildLogic) {
-            settingsGradleKts {
-                """
-                pluginManagement {
-                    repositories {
-                        gradlePluginPortal()
-                        mavenLocal()
-                    }
-                }
-                    
-                plugins {
-                    id("dev.panuszewski.typesafe-conventions") version "$projectVersion"
-                }
-                """
-            }
-
-            buildGradleKts {
-                """
-                plugins {
-                    `kotlin-dsl`
-                } 
-                
-                repositories {
-                    mavenCentral()
-                }
-                """
-            }
-
-            customProjectFile("src/main/kotlin/someConvention.gradle.kts") {
-                """
-                plugins {
-                    java
-                }
-                
-                println("Hello from someConvention")
-                
-                dependencies {
-                    implementation(libs.some.library)
-                }    
-                """
-            }
-        }
+        installFixture(ConventionCatalogUsedInParentBuildThatIsNotRootBuild)
 
         // when
         val result = runGradle()
@@ -497,53 +388,7 @@ class ConventionPluginsSpec : GradleSpec() {
     @Test
     fun `should not add convention catalog to root build in multi-level hierarchy`() {
         // given
-        val secondaryBuild = mainBuild.registerIncludedBuild("secondary-build")
-        val buildLogic = secondaryBuild.registerIncludedBuild("build-logic")
-
-        with(mainBuild) {
-            buildGradleKts {
-                """
-                plugins {
-                    alias(conventions.plugins.someConvention)
-                }
-                """
-            }
-        }
-
-        with(buildLogic) {
-            settingsGradleKts {
-                """
-                pluginManagement {
-                    repositories {
-                        gradlePluginPortal()
-                        mavenLocal()
-                    }
-                }
-                    
-                plugins {
-                    id("dev.panuszewski.typesafe-conventions") version "$projectVersion"
-                }
-                """
-            }
-
-            buildGradleKts {
-                """
-                plugins {
-                    `kotlin-dsl`
-                } 
-                
-                repositories {
-                    mavenCentral()
-                }
-                """
-            }
-
-            customProjectFile("src/main/kotlin/someConvention.gradle.kts") {
-                """
-                println("Hello from someConvention")
-                """
-            }
-        }
+        installFixture(ConventionCatalogUsedInRootBuildThatIsNotDirectParent)
 
         // when
         val result = runGradle()
@@ -557,53 +402,7 @@ class ConventionPluginsSpec : GradleSpec() {
     fun `should generate typesafe accessor for convention plugin and use it from another convention plugin`() {
         // given
         installFixture(BuildLogic)
-        installFixture(TypesafeConventionsAppliedToIncludedBuild)
-
-        buildGradleKts {
-            """
-            plugins {
-                alias(conventions.plugins.someConvention)
-            }
-            
-            repositories {
-                mavenCentral()
-            }
-            """
-        }
-
-        includedBuild {
-            buildGradleKts {
-                """
-                plugins {
-                    `kotlin-dsl`
-                }
-                
-                repositories {
-                    gradlePluginPortal()
-                }
-                """
-            }
-
-            customProjectFile("src/main/kotlin/conventions/someConvention.gradle.kts") {
-                """
-                package conventions
-                    
-                import conventions
-                
-                plugins {
-                    alias(conventions.plugins.anotherConvention)
-                }    
-                """
-            }
-
-            customProjectFile("src/main/kotlin/conventions/anotherConvention.gradle.kts") {
-                """
-                package conventions
-                
-                println("Hello from anotherConvention")
-                """
-            }
-        }
+        installFixture(ConventionCatalogUsedFromConventionPlugin)
 
         // when
         val result = runGradle()
