+++ src/test/kotlin/dev/panuszewski/gradle/fixtures/EarlyEvaluatedIncludedBuild.kt
@@ -9,19 +9,6 @@ object EarlyEvaluatedIncludedBuild : NoConfigFixture {
     override fun GradleSpec.install() {
         installFixture(PluginManagementBuildLogic)
         installFixture(TypesafeConventionsAppliedToIncludedBuild)
-
-        includedBuild {
-            buildGradleKts {
-                """
-                plugins {
-                    `kotlin-dsl`    
-                }
-                
-                repositories {
-                    mavenCentral()
-                }
-                """
-            }
-        }
+        installFixture(IncludedBuildConfiguredForHostingConventions)
     }
 }
