+++ src/test/kotlin/dev/panuszewski/gradle/fixtures/IncludedBuildConfiguredForHostingConventions.kt
@@ -1,27 +1,24 @@
 package dev.panuszewski.gradle.fixtures
 
-import dev.panuszewski.gradle.fixtures.TypesafeConventionsAppliedToIncludedBuild.Config
+import dev.panuszewski.gradle.fixtures.IncludedBuildConfiguredForHostingConventions.Config
 import dev.panuszewski.gradle.framework.Fixture
 import dev.panuszewski.gradle.framework.GradleBuild
 import dev.panuszewski.gradle.framework.GradleSpec
 
-object TypesafeConventionsAppliedToIncludedBuild : Fixture<Config> {
+object IncludedBuildConfiguredForHostingConventions : Fixture<Config> {
 
     override fun GradleSpec.install(config: Config) {
         val build = config.build ?: singleIncludedBuild()
 
         with(build) {
-            settingsGradleKts {
+            buildGradleKts {
                 """
-                pluginManagement {
-                    repositories {
-                        gradlePluginPortal()
-                        mavenLocal()
-                    }
-                }
-                    
                 plugins {
-                    id("dev.panuszewski.typesafe-conventions") version "$projectVersion"
+                    `kotlin-dsl`
+                }
+                
+                repositories {
+                    gradlePluginPortal()
                 }
                 """
             }
